cotizacion etapa 4

+++ routes/web.php
@@ -8,6 +8,11 @@ use App\Http\Controllers\EmpresaController;
 use App\Http\Controllers\ProgramaController;
 use App\Http\Controllers\OfertaController;
 use App\Http\Controllers\CotizacionController;
+use App\Http\Controllers\PdfController;
+use App\Models\User;
+use App\Models\Cotizacion;
+use App\Models\Empresa;
+
 
 
 
@@ -22,8 +27,10 @@ use App\Http\Controllers\CotizacionController;
 |
 */
 
+// Route::get('/', [LoginController::class, 'actions']);
+
 Route::get('/', function () {
-    return view('login');
+    return view('welcome');
 });
 
 Route::get('/home', [LoginController::class, 'index'])->middleware('auth')->name('home');
@@ -31,37 +38,50 @@ Route::get('/home', [LoginController::class, 'index'])->middleware('auth')->name
 Route::get('/registro', [RegistroController::class, 'index']);
 // Route::get('/register', [RegistroController::class, 'index']);
 
-Route::get('/usuarios/crear', [UserController::class, 'create'])->middleware('auth')->name('trabajador.registro');
+Route::get('/usuarios/crear', [UserController::class, 'create'])->middleware('auth','admin')->name('trabajador.registro');
 
-Route::post('usuarios/crear', [UserController::class, 'store'])->middleware('auth')->name('trabajador.guardar');
+Route::post('usuarios/crear', [UserController::class, 'store'])->middleware('auth', 'admin')->name('trabajador.guardar');
 
-Route::get('usuarios/listado', [UserController::class , 'index'])->middleware('auth')->name('usuarios.listado');
+Route::get('usuarios/listado', [UserController::class , 'index'])->middleware('auth','adminworker')->name('usuarios.listado');
 
-Route::get('usuarios/{id}', [UserController::class , 'show'])->middleware('auth')->name('usuarios.detalle');
+Route::get('usuarios/{id}', [UserController::class , 'show'])->middleware('auth','thisuser')->name('usuarios.detalle');
 
-Route::get('usuarios/{id}/editar', [UserController::class , 'edit'])->middleware('auth')->name('usuarios.editar');
-Route::put('usuarios/{id}', [UserController::class , 'update'])->middleware('auth')->name('usuarios.actualizar');
-Route::put('usuarios/{id}/cambiarcontraseña', [UserController::class , 'updatePassword'])->middleware('auth')->name('usuarios.actualizarcontraseña');
-Route::put('usuarios/{id}/empresa/editar', [EmpresaController::class , 'update'])->middleware('auth')->name('empresas.actualizar');
+Route::get('usuarios/{id}/editar', [UserController::class , 'edit'])->middleware('auth', 'thisuser')->name('usuarios.editar');
+Route::put('usuarios/{id}', [UserController::class , 'update'])->middleware('auth', 'thisuser')->name('usuarios.actualizar');
+Route::put('usuarios/{id}/cambiarcontraseña', [UserController::class , 'updatePassword'])->middleware('auth', 'thisuser')->name('usuarios.actualizarcontraseña');
+Route::put('usuarios/{id}/empresa/editar', [EmpresaController::class , 'update'])->middleware('auth' , 'thisuser')->name('empresas.actualizar');
 
 Route::get('usuarios/{id}/empresa', [EmpresaController::class , 'edit'])->middleware('auth')->name('empresas.editar');
 Route::delete('usuarios/{id}/eliminar', [UserController::class , 'destroy'])->middleware('auth')->name('usuarios.eliminar');
 
 Route::get('programas/listado', [ProgramaController::class , 'index'])->middleware('auth')->name('programas.listado');
-Route::get('programas/crear', [ProgramaController::class , 'create'])->middleware('auth')->name('programas.registro');
-Route::post('programas/crear', [ProgramaController::class , 'store'])->middleware('auth')->name('programas.guardar');
-Route::get('programas/{id}/editar', [ProgramaController::class , 'edit'])->middleware('auth')->name('programas.editar');
-Route::put('programas/{id}', [ProgramaController::class , 'update'])->middleware('auth')->name('programas.actualizar');
-Route::delete('programas/{id}/eliminar', [ProgramaController::class , 'destroy'])->middleware('auth')->name('programas.eliminar');
+Route::get('programas/crear', [ProgramaController::class , 'create'])->middleware('auth', 'admin')->name('programas.registro');
+Route::post('programas/crear', [ProgramaController::class , 'store'])->middleware('auth','admin')->name('programas.guardar');
+Route::get('programas/{id}/editar', [ProgramaController::class , 'edit'])->middleware('auth', 'admin')->name('programas.editar');
+Route::put('programas/{id}', [ProgramaController::class , 'update'])->middleware('auth','admin')->name('programas.actualizar');
+Route::delete('programas/{id}/eliminar', [ProgramaController::class , 'destroy'])->middleware('auth','admin')->name('programas.eliminar');
 
-Route::get('ofertas/listado', [OfertaController::class , 'index'])->middleware('auth')->name('ofertas.listado');
-Route::get('ofertas/crear', [OfertaController::class , 'create'])->middleware('auth')->name('ofertas.registro');
-Route::post('ofertas/crear', [OfertaController::class , 'store'])->middleware('auth')->name('ofertas.guardar');
-Route::delete('ofertas/{id}/eliminar', [OfertaController::class , 'destroy'])->middleware('auth')->name('ofertas.eliminar');
+Route::get('ofertas/listado', [OfertaController::class , 'index'])->middleware('auth','adminworker')->name('ofertas.listado');
+Route::get('ofertas/crear', [OfertaController::class , 'create'])->middleware('auth', 'admin')->name('ofertas.registro');
+Route::post('ofertas/crear', [OfertaController::class , 'store'])->middleware('auth','admin')->name('ofertas.guardar');
+Route::delete('ofertas/{id}/eliminar', [OfertaController::class , 'destroy'])->middleware('auth','admin')->name('ofertas.eliminar');
 
 
-Route::get('cotizaciones/crear', [CotizacionController::class , 'create'])->middleware('auth')->name('cotizaciones.registro');
-Route::post('cotizaciones/crear', [CotizacionController::class , 'store'])->middleware('auth')->name('cotizaciones.guardar');
+Route::get('cotizaciones/crear', [CotizacionController::class , 'create'])->middleware('auth', 'user')->name('cotizaciones.registro');
+Route::post('cotizaciones/crear', [CotizacionController::class , 'store'])->middleware('auth', 'user')->name('cotizaciones.guardar');
+Route::get('cotizaciones/responder/{id}', [CotizacionController::class , 'create2'])->middleware('auth')->name('cotizaciones.responder');
+Route::post('cotizaciones/responder/{id}', [CotizacionController::class , 'store2'])->middleware('auth')->name('cotizaciones.guardar2');
+Route::get('cotizaciones/listado', [CotizacionController::class , 'index'])->middleware('auth')->name('cotizaciones.listado');
+
+
+Route::get('pdf/{id}', [PdfController::class , 'showPdf'])->middleware('auth', 'adminworker')->name('pdf.mostrar');
+
+
+    
+    
+    
+
+
 
 
 // Route::get('/home', function () {
